+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
+* Fixed issue with default outfile file name for **unify** command when using an input of ".".
 * *Unify* command now adds contributor details to the display name and description.
 * *Format* command now adds *isFetch* and *incidenttype* fields to integration yml.
 * Removed the *feedIncremental* field from the integration schema.
 * *Format* command now adds *feedBypassExclusionList*, *Fetch indicators*, *feedReputation*, *feedReliability*,
      *feedExpirationPolicy*, *feedExpirationInterval* and *feedFetchInterval* fields to integration yml.
+* Fixed an issue in the playbooks schema.
+* Fixed an issue where generated release notes were out of order.
+* Improved pack dependencies detection.
 
 #### 1.1.3
 * Added a validation for invalid id fields in indicators types files in **validate** command.